tegra20: add generic board and Colibri T20 on Iris
The Tegra arch will be fully based on device tree and most boards
shouldn't need anything more than the generic drivers and arch code.

This adds generic tegra20 board support, but since we currently have
no way to combine an image with a devicetree other than compiling it
into the single image we also add Colibri T20 on Iris support here
and add a defconfig for it.

Signed-off-by: Lucas Stach <dev@lynxeye.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>