ARM: Makefile: Make sure to clean boarddirs
On ARM the boarddirs are only cleaned due to the regular file
pattern search, but are not explicitly cleaned. This makes
it impossible to clean files which are not matched by a pattern
but have to be explicitly cleaned due to adding them to extra-y.

This patch changes the board-/board-y variables to += so that
we can use $(board-) to add it to common-. This way the board
directories are also cleaned.

While touching the board-y variables anyway order them
alphabetically.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>